Move parseRawPermissionsJson function to composable
